Finish moving remaining components and resources to modules
