+++ Modules/Avatar/app/View/Components/CampaignOptions.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\View\Components\Cyberpunkred;
+namespace Modules\Avatar\View\Components;
 
 use Illuminate\Contracts\View\View;
 use Illuminate\View\Component;
@@ -14,6 +14,6 @@ class CampaignOptions extends Component
 {
     public function render(): View
     {
-        return view('components.cyberpunkred.campaign-options');
+        return view('avatar::components.campaign-options');
     }
 }
