+++ Modules/Subversion/app/View/Components/CampaignOptions.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Modules\Subversion\View\Components;
+
+use Illuminate\Contracts\View\View;
+use Illuminate\View\Component;
+
+/**
+ * @psalm-suppress UnusedClass
+ */
+class CampaignOptions extends Component
+{
+    public function render(): View
+    {
+        return view('subversion::components.campaign-options');
+    }
+}
